+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * 【poi    】      增加GlobalPoiConfig（issue#IAEHJH@Gitee）
 * 【core   】      优化IndexedComparator性能（pr#1240@Gitee）
 * 【http   】      改进ContentType.get忽略空格（pr#3664@Github）
+* 【http   】      CompressUtil.createExtractor支持tgz自动识别（pr#3674@Github）
 
 ### 🐞Bug修复
 * 【core   】      修复因RFC3986理解有误导致的UrlPath处理冒号转义问题（issue#IAAE88@Gitee）